Begin refactor by swapping to ts
