+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-chartlet",
   "version": "1.2.0",
   "description": "A dead simple and tiny charting library for React",
-  "source": "src/index.js",
+  "source": "src/index.ts",
   "main": "dist/main.js",
   "module": "dist/module.js",
   "author": "Stevent team (https://stevent.dev)",
