+++ src/components/ChartSVG.tsx
@@ -1,0 +1,23 @@
+const ChartSVG: React.FC<React.SVGProps<SVGSVGElement>> = ({
+  width,
+  height,
+  ...props
+}) => {
+  return (
+    <svg
+      width={width}
+      height={height}
+      viewBox={`0 0 ${width} ${height}`}
+      {...props}
+      style={{
+        display: 'block',
+        boxSizing: 'border-box',
+        position: 'absolute',
+        inset: 0,
+        ...props.style
+      }}
+    />
+  )
+}
+
+export default ChartSVG
