--- src/index.js
@@ -1,3 +1,0 @@
-export { default as Sparkline } from './charts/Sparkline'
-export { default as Donut } from './charts/Donut'
-export { default as Legend } from './components/Legend'
