+++ src/index.ts
@@ -1,0 +1,2 @@
+export { default as DonutChart } from './charts/DonutChart'
+export { default as Legend } from './components/Legend'
